+++ src/DotNetWorker.Core/Http/HttpResponseDataExtensions.cs
@@ -81,7 +81,7 @@ namespace Microsoft.Azure.Functions.Worker.Http
 
         /// <summary>
         /// Asynchronously writes the specified value as JSON to the response body using the default <see cref="ObjectSerializer"/> configured for this worker.
-        /// The response content-type will be set to <code>application/json; charset=utf-8</code> and the status code set to 200.
+        /// The response content-type will be set to <code>application/json; charset=utf-8</code>.
         /// </summary>
         /// <typeparam name="T">The type of object to write.</typeparam>
         /// <param name="response">The response to write JSON to.</param>
@@ -90,28 +90,12 @@ namespace Microsoft.Azure.Functions.Worker.Http
         /// <returns>A <see cref="ValueTask"/> that represents the asynchronous operation.</returns>
         public static ValueTask WriteAsJsonAsync<T>(this HttpResponseData response, T instance, CancellationToken cancellationToken = default)
         {
-            return WriteAsJsonAsync(response, instance, "application/json; charset=utf-8", HttpStatusCode.OK, cancellationToken);
+            return WriteAsJsonAsync(response, instance, "application/json; charset=utf-8", cancellationToken);
         }
 
         /// <summary>
         /// Asynchronously writes the specified value as JSON to the response body using the default <see cref="ObjectSerializer"/> configured for this worker.
-        /// The response content-type will be set to <code>application/json; charset=utf-8</code> and the status code set to the provided <paramref name="statusCode"/>.
-        /// </summary>
-        /// <typeparam name="T">The type of object to write.</typeparam>
-        /// <param name="response">The response to write JSON to.</param>
-        /// <param name="instance">The instance to serialize and write as JSON.</param>
-        /// <param name="statusCode">The status code to set on the response.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> used to cancel the operation.</param>
-        /// <returns>A <see cref="ValueTask"/> that represents the asynchronous operation.</returns>
-        public static ValueTask WriteAsJsonAsync<T>(this HttpResponseData response, T instance, HttpStatusCode statusCode,
-            CancellationToken cancellationToken = default)
-        {
-            return WriteAsJsonAsync(response, instance, "application/json; charset=utf-8", statusCode, cancellationToken);
-        }
-
-        /// <summary>
-        /// Asynchronously writes the specified value as JSON to the response body using the default <see cref="ObjectSerializer"/> configured for this worker.
-        /// The response content-type will be set to the provided <paramref name="contentType"/> and the status code set to 200.
+        /// The response content-type will be set to the provided <paramref name="contentType"/>.
         /// </summary>
         /// <typeparam name="T">The type of object to write.</typeparam>
         /// <param name="response">The response to write JSON to.</param>
@@ -128,37 +112,12 @@ namespace Microsoft.Azure.Functions.Worker.Http
 
             ObjectSerializer serializer = GetObjectSerializer(response);
 
-            return WriteAsJsonAsync(response, instance, serializer, contentType, HttpStatusCode.OK, cancellationToken);
+            return WriteAsJsonAsync(response, instance, serializer, contentType, cancellationToken);
         }
-
-        /// <summary>
-        /// Asynchronously writes the specified value as JSON to the response body using the default <see cref="ObjectSerializer"/> configured for this worker.
-        /// The response content-type will be set to the provided <paramref name="contentType"/> and the status code set to the provided <paramref name="statusCode"/>.
-        /// </summary>
-        /// <typeparam name="T">The type of object to write.</typeparam>
-        /// <param name="response">The response to write JSON to.</param>
-        /// <param name="instance">The instance to serialize and write as JSON.</param>
-        /// <param name="contentType">The content-type to set on the response.</param>
-        /// <param name="statusCode">The status code to set on the response.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> used to cancel the operation.</param>
-        /// <returns>A <see cref="ValueTask"/> that represents the asynchronous operation.</returns>
-        public static ValueTask WriteAsJsonAsync<T>(this HttpResponseData response, T instance, string contentType, HttpStatusCode statusCode,
-            CancellationToken cancellationToken = default)
-        {
-            if (response is null)
-            {
-                throw new ArgumentNullException(nameof(response));
-            }
-
-            ObjectSerializer serializer = GetObjectSerializer(response);
-
-            return WriteAsJsonAsync(response, instance, serializer, contentType, statusCode, cancellationToken);
-        }
-
 
         /// <summary>
         /// Asynchronously writes the specified value as JSON to the response body using the provided <see cref="ObjectSerializer"/>.
-        /// The response content-type will be set to <code>application/json; charset=utf-8</code> and the status code set to 200.
+        /// The response content-type will be set to <code>application/json; charset=utf-8</code>.
         /// </summary>
         /// <typeparam name="T">The type of object to write.</typeparam>
         /// <param name="response">The response to write JSON to.</param>
@@ -168,29 +127,12 @@ namespace Microsoft.Azure.Functions.Worker.Http
         /// <returns>A <see cref="ValueTask"/> that represents the asynchronous operation.</returns>
         public static ValueTask WriteAsJsonAsync<T>(this HttpResponseData response, T instance, ObjectSerializer serializer, CancellationToken cancellationToken = default)
         {
-            return WriteAsJsonAsync(response, instance, serializer, "application/json; charset=utf-8", HttpStatusCode.OK, cancellationToken);
+            return WriteAsJsonAsync(response, instance, serializer, "application/json; charset=utf-8", cancellationToken);
         }
-
+                
         /// <summary>
         /// Asynchronously writes the specified value as JSON to the response body using the provided <see cref="ObjectSerializer"/>.
-        /// The response content-type will be set to <code>application/json; charset=utf-8</code> and the status code set to the provided <paramref name="statusCode"/>.
-        /// </summary>
-        /// <typeparam name="T">The type of object to write.</typeparam>
-        /// <param name="response">The response to write JSON to.</param>
-        /// <param name="instance">The instance to serialize and write as JSON.</param>
-        /// <param name="serializer">The serializer used to serialize the instance.</param>
-        /// <param name="statusCode">The status code to set on the response.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> used to cancel the operation.</param>
-        /// <returns>A <see cref="ValueTask"/> that represents the asynchronous operation.</returns>
-        public static ValueTask WriteAsJsonAsync<T>(this HttpResponseData response, T instance, ObjectSerializer serializer, HttpStatusCode statusCode,
-            CancellationToken cancellationToken = default)
-        {
-            return WriteAsJsonAsync(response, instance, serializer, "application/json; charset=utf-8", statusCode, cancellationToken);
-        }
-
-        /// <summary>
-        /// Asynchronously writes the specified value as JSON to the response body using the provided <see cref="ObjectSerializer"/>.
-        /// The response content-type will be set to the provided <paramref name="contentType"/> and the status code set to 200.
+        /// The response content-type will be set to the provided <paramref name="contentType"/>.
         /// </summary>
         /// <typeparam name="T">The type of object to write.</typeparam>
         /// <param name="response">The response to write JSON to.</param>
@@ -199,26 +141,7 @@ namespace Microsoft.Azure.Functions.Worker.Http
         /// <param name="contentType">The content-type to set on the response.</param>
         /// <param name="cancellationToken">A <see cref="CancellationToken"/> used to cancel the operation.</param>
         /// <returns>A <see cref="ValueTask"/> that represents the asynchronous operation.</returns>
-        public static ValueTask WriteAsJsonAsync<T>(this HttpResponseData response, T instance,
-            ObjectSerializer serializer, string contentType,
-            CancellationToken cancellationToken = default)
-        {
-            return WriteAsJsonAsync(response, instance, serializer, contentType, HttpStatusCode.OK, cancellationToken);
-        }
-
-        /// <summary>
-        /// Asynchronously writes the specified value as JSON to the response body using the provided <see cref="ObjectSerializer"/>.
-        /// The response content-type will be set to the provided <paramref name="contentType"/> and the status code set to the provided <paramref name="statusCode"/>.
-        /// </summary>
-        /// <typeparam name="T">The type of object to write.</typeparam>
-        /// <param name="response">The response to write JSON to.</param>
-        /// <param name="instance">The instance to serialize and write as JSON.</param>
-        /// <param name="serializer">The serializer used to serialize the instance.</param>
-        /// <param name="contentType">The content-type to set on the response.</param>
-        /// <param name="statusCode">The status code to set on the response.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> used to cancel the operation.</param>
-        /// <returns>A <see cref="ValueTask"/> that represents the asynchronous operation.</returns>
-        public static ValueTask WriteAsJsonAsync<T>(this HttpResponseData response, T instance, ObjectSerializer serializer, string contentType, HttpStatusCode statusCode, CancellationToken cancellationToken = default)
+        public static ValueTask WriteAsJsonAsync<T>(this HttpResponseData response, T instance, ObjectSerializer serializer, string contentType, CancellationToken cancellationToken = default)
         {
             if (response is null)
             {
@@ -236,8 +159,6 @@ namespace Microsoft.Azure.Functions.Worker.Http
             }
 
             response.Headers.Add("Content-Type", contentType);
-            response.StatusCode = statusCode;
-
             return serializer.SerializeAsync(response.Body, instance, typeof(T), cancellationToken);
         }